@@ -1,6 +1,6 @@
 name: swagger_dart_code_generator
 
-version: 2.0.0-nullsafety.2
+version: 2.0.0-nullsafety.3+3
 
 homepage: https://github.com/epam-cross-platform-lab/swagger-dart-code-generator
 repository: https://github.com/epam-cross-platform-lab/swagger-dart-code-generator
@@ -11,26 +11,21 @@ description: Have you been turned into a problem with writing code for Http
   functionality you have been looking for.
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  build: ^2.0.0
+  build: ^2.0.1
   path: ^1.8.0
   markdown: ^4.0.0
-  recase: ^4.0.0-nullsafety.0
-  json_annotation: ^4.0.0
-  build_runner: ^1.11.1
+  recase: ^4.0.0
+  json_annotation: ^4.0.1
   meta: ^1.3.0
-  build_config: ^0.4.5
-  universal_io: ^2.0.0-nullsafety.2
+  universal_io: ^2.0.4
+  collection: ^1.15.0
   dart_style: ^2.0.0
 
 dev_dependencies:
-  json_serializable: ^4.0.0
-  flutter_test:
-    sdk: flutter
-  pedantic: ^1.10.0
-  test: any
-
-dependency_overrides:
-  analyzer: ^1.0.0
+  json_serializable: ^4.1.1
+  pedantic: ^1.11.0
+  test: ^1.16.8
+ 
