@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:swagger_dart_code_generator/src/models/generator_options.dart';
 import 'package:recase/recase.dart';
 import 'package:swagger_dart_code_generator/src/code_generators/v2/swagger_enums_generator_v2.dart';
@@ -67,6 +69,40 @@ abstract class SwaggerModelsGenerator {
     );
   }
 
+  static Map<String, dynamic> getClassesFromResponses(String dartCode) {
+    final swagger = jsonDecode(dartCode);
+
+    final results = <String, dynamic>{};
+
+    final paths = swagger['paths'] as Map<String, dynamic>?;
+    if (paths == null) {
+      return results;
+    }
+
+    paths.forEach((path, pathValue) {
+      final requests = pathValue as Map<String, dynamic>;
+
+      requests.removeWhere((key, value) => key == 'parameters');
+
+      requests.forEach((request, requestValue) {
+        final responses = requestValue['responses'] as Map<String, dynamic>;
+
+        final neededResponse = responses['200'];
+
+        if (neededResponse != null &&
+            neededResponse['schema'] != null &&
+            neededResponse['schema']['type'] == 'object' &&
+            neededResponse['schema']['properties'] != null) {
+          final pathText = path.split('/').map((e) => e.pascalCase).join();
+          final requestText = request.pascalCase;
+          results['$pathText$requestText\$Response'] = neededResponse['schema'];
+        }
+      });
+    });
+
+    return results;
+  }
+
   String generateBase(
       String dartCode,
       String fileName,
@@ -81,7 +117,11 @@ abstract class SwaggerModelsGenerator {
             allEnumsNames, options.enumsCaseSensitive)
         : '';
 
-    if (classes == null) {
+    final classesFromResponses = getClassesFromResponses(dartCode);
+
+    classes.addAll(classesFromResponses);
+
+    if (classes.isEmpty) {
       return '';
     }
 
@@ -89,6 +129,7 @@ abstract class SwaggerModelsGenerator {
       if (classes['enum'] != null) {
         return '';
       }
+
       return generateModelClassContent(
         className.pascalCase,
         classes[className] as Map<String, dynamic>,
@@ -113,7 +154,7 @@ abstract class SwaggerModelsGenerator {
   }
 
   static String getValidatedClassName(String className) {
-    if (className == null) {
+    if (className.isEmpty) {
       return className;
     }
 
@@ -141,12 +182,12 @@ abstract class SwaggerModelsGenerator {
 
   String getParameterTypeName(
       String className, String parameterName, Map<String, dynamic> parameter,
-      [String refNameParameter]) {
+      [String? refNameParameter]) {
     if (refNameParameter != null) {
       return refNameParameter.pascalCase;
     }
 
-    if (parameter == null) {
+    if (parameter.isEmpty) {
       return 'Object';
     }
 
@@ -154,7 +195,7 @@ abstract class SwaggerModelsGenerator {
       return parameter['\$ref'].toString().split('/').last.pascalCase;
     }
 
-    switch (parameter['type'] as String) {
+    switch (parameter['type'] as String?) {
       case 'integer':
       case 'int':
         return 'int';
@@ -173,7 +214,7 @@ abstract class SwaggerModelsGenerator {
       case 'object':
         return 'Object';
       case 'array':
-        final items = parameter['items'] as Map<String, dynamic>;
+        final items = parameter['items'] as Map<String, dynamic>? ?? {};
         return getParameterTypeName(className, parameterName, items);
       default:
         if (parameter['oneOf'] != null) {
@@ -243,11 +284,11 @@ abstract class SwaggerModelsGenerator {
   }
 
   String generateIncludeIfNullString(GeneratorOptions options) {
-    if (options.includeIfNull == null || !options.includeIfNull.enabled) {
+    if (options.includeIfNull == null || !options.includeIfNull!.enabled) {
       return '';
     }
 
-    return ', includeIfNull: ${options.includeIfNull.value}';
+    return ', includeIfNull: ${options.includeIfNull!.value}';
   }
 
   String generatePropertyContentByDefault(
@@ -256,9 +297,12 @@ abstract class SwaggerModelsGenerator {
       List<String> allEnumNames,
       List<String> allEnumListNames,
       GeneratorOptions options) {
-    final typeName =
-        getValidatedClassName(propertyEntryMap['originalRef'].toString()) ??
-            'dynamic';
+    var typeName =
+        getValidatedClassName(propertyEntryMap['originalRef'].toString());
+
+    if (typeName.isEmpty) {
+      typeName = 'dynamic';
+    }
 
     final unknownEnumValue = generateUnknownEnumValue(
         allEnumNames, allEnumListNames, typeName.toString(), false);
@@ -294,8 +338,8 @@ abstract class SwaggerModelsGenerator {
   }
 
   String generateToJsonForDate(Map<String, dynamic> map) {
-    final type = map['type']?.toString()?.toLowerCase();
-    final format = map['format']?.toString()?.toLowerCase();
+    final type = map['type']?.toString().toLowerCase();
+    final format = map['format']?.toString().toLowerCase();
 
     final isDate = type == 'string' && format == 'date';
 
@@ -320,7 +364,7 @@ abstract class SwaggerModelsGenerator {
 
     String typeName;
     if (basicTypesMap.containsKey(parameterName)) {
-      typeName = basicTypesMap[parameterName];
+      typeName = basicTypesMap[parameterName]!;
     } else {
       typeName = getValidatedClassName(getParameterTypeName(
           className, propertyName, propertyEntryMap, parameterName));
@@ -358,7 +402,7 @@ abstract class SwaggerModelsGenerator {
     final parameterName = propertyEntryMap['\$ref'].toString().split('/').last;
     String typeName;
     if (basicTypesMap.containsKey(parameterName)) {
-      typeName = basicTypesMap[parameterName];
+      typeName = basicTypesMap[parameterName]!;
     } else {
       typeName = getValidatedClassName(getParameterTypeName(
           className, propertyName, propertyEntryMap, parameterName));
@@ -399,7 +443,8 @@ abstract class SwaggerModelsGenerator {
     allEnumNames.add(enumName);
 
     final unknownEnumValue = generateUnknownEnumValue(
-        allEnumNames, allEnumListNames, enumName, false);
+            allEnumNames, allEnumListNames, enumName, false)
+        .substring(2);
 
     final includeIfNullString = generateIncludeIfNullString(options);
 
@@ -418,20 +463,20 @@ abstract class SwaggerModelsGenerator {
       List<String> allEnumListNames,
       GeneratorOptions options,
       Map<String, String> basicTypesMap) {
-    final dynamic items = propertyEntryMap['items'];
+    final items = propertyEntryMap['items'];
 
-    String typeName;
+    var typeName = '';
     if (items != null) {
-      typeName = getValidatedClassName(items['originalRef'] as String);
+      typeName = getValidatedClassName(items['originalRef'] as String? ?? '');
 
-      if (typeName == null) {
-        final ref = items['\$ref'] as String;
-        typeName = ref?.split('/')?.last;
+      if (typeName.isEmpty) {
+        final ref = items['\$ref'] as String?;
+        typeName = ref?.split('/').last ?? 'dynamic';
 
         if (basicTypesMap.containsKey(typeName)) {
-          typeName = basicTypesMap[typeName];
-        } else {
-          typeName = typeName?.pascalCase;
+          typeName = basicTypesMap[typeName]!;
+        } else if (typeName != 'dynamic') {
+          typeName = typeName.pascalCase;
         }
       }
 
@@ -440,8 +485,10 @@ abstract class SwaggerModelsGenerator {
       }
     }
 
-    typeName ??= getParameterTypeName(
-        className, propertyName, items as Map<String, dynamic>);
+    if (typeName.isEmpty) {
+      typeName = getParameterTypeName(
+          className, propertyName, items as Map<String, dynamic>? ?? {});
+    }
 
     final unknownEnumValue = generateUnknownEnumValue(
         allEnumNames, allEnumListNames, typeName, true);
@@ -530,7 +577,6 @@ abstract class SwaggerModelsGenerator {
           options,
           basicTypesMap,
         );
-        break;
       case 'enum':
         return generateEnumPropertyContent(
           propertyName,
@@ -539,7 +585,6 @@ abstract class SwaggerModelsGenerator {
           allEnumListNames,
           options,
         );
-        break;
       default:
         return generateGeneralPropertyContent(
           propertyName,
@@ -563,7 +608,7 @@ abstract class SwaggerModelsGenerator {
       List<String> allEnumNames,
       List<String> allEnumListNames,
       GeneratorOptions options) {
-    if (propertiesMap == null) {
+    if (propertiesMap.isEmpty) {
       return '';
     }
 
@@ -627,7 +672,7 @@ abstract class SwaggerModelsGenerator {
   static Map<String, String> generateBasicTypesMapFromSchemas(
       Map<String, dynamic> schemas) {
     final result = <String, String>{};
-    if (schemas == null) {
+    if (schemas.isEmpty) {
       return result;
     }
 
@@ -663,7 +708,7 @@ abstract class SwaggerModelsGenerator {
       case 'boolean':
         return 'bool';
       default:
-        return null;
+        return '';
     }
   }
 
@@ -734,7 +779,7 @@ List<enums.$neededName> ${neededName.camelCase}ListFromJson(
     List<String> allEnumNames,
     List<String> allEnumListNames,
   ) {
-    if (entityMap == null) {
+    if (entityMap.isEmpty) {
       return '';
     }
 
@@ -746,13 +791,12 @@ List<enums.$neededName> ${neededName.camelCase}ListFromJson(
       final hasDefaultValue = value['default'] != null ||
           defaultValues.any((element) =>
               element.typeName ==
-              _mapBasicTypeToDartType(value['type'].toString(), null));
+              _mapBasicTypeToDartType(value['type'].toString(), ''));
 
       final isList = value['type'] == 'array' ||
           allEnumListNames.contains('enums.${key.pascalCase}');
 
-      final type =
-          value['\$ref']?.toString()?.split('/')?.last?.pascalCase ?? key;
+      final type = value['\$ref']?.toString().split('/').last.pascalCase ?? key;
 
       final isEnum = allEnumNames.contains('enums.${type.pascalCase}') ||
           allEnumNames.contains('enums.${className + type.pascalCase}');
