@@ -10,6 +10,7 @@ import 'package:swagger_dart_code_generator/src/swagger_models/swagger_path.dart
 import 'package:swagger_dart_code_generator/src/swagger_models/swagger_root.dart';
 import 'package:recase/recase.dart';
 import 'package:swagger_dart_code_generator/src/exception_words.dart';
+import 'package:collection/collection.dart';
 
 abstract class SwaggerRequestsGenerator {
   static const String defaultBodyParameter = 'Object';
@@ -71,10 +72,10 @@ $allMethodsContent
   static List<String> getAllDynamicResponses(String dartCode) {
     final dynamic map = jsonDecode(dartCode);
 
-    final components = map['components'] as Map<String, dynamic>;
+    final components = map['components'] as Map<String, dynamic>?;
     final responses = components == null
         ? null
-        : components['responses'] as Map<String, dynamic>;
+        : components['responses'] as Map<String, dynamic>?;
 
     if (responses == null) {
       return [];
@@ -83,10 +84,11 @@ $allMethodsContent
     var results = <String>[];
 
     responses.keys.forEach((key) {
-      final response = responses[key] as Map<String, dynamic>;
+      final response = responses[key] as Map<String, dynamic>?;
 
-      final content =
-          response == null ? null : response['content'] as Map<String, dynamic>;
+      final content = response == null
+          ? null
+          : response['content'] as Map<String, dynamic>?;
 
       if (content != null && content.entries.length > 1) {
         results.add(key.capitalize);
@@ -106,8 +108,8 @@ $allMethodsContent
   ) {
     final methods = StringBuffer();
 
-    final dynamic map = jsonDecode(dartCode);
-    final components = map['components'] as Map<String, dynamic>;
+    final dynamic map = dartCode.isNotEmpty ? jsonDecode(dartCode) : {};
+    final components = map['components'] as Map<String, dynamic>?;
     final requestBodies =
         components == null ? null : components['requestBodies'];
 
@@ -117,7 +119,7 @@ $allMethodsContent
               swaggerRequest.type.toLowerCase() != requestTypeOptions)
           .forEach((SwaggerRequest swaggerRequest) {
         swaggerRequest.parameters = swaggerRequest.parameters
-            .where((element) => element?.inParameter != null)
+            .where((element) => element.inParameter.isNotEmpty)
             .toList();
 
         final hasFormData = swaggerRequest.parameters.any(
@@ -126,7 +128,7 @@ $allMethodsContent
 
         String methodName;
         if (options.usePathForRequestNames ||
-            swaggerRequest.operationId == null) {
+            swaggerRequest.operationId.isEmpty) {
           methodName = SwaggerModelsGenerator.generateRequestName(
               swaggerPath.path, swaggerRequest.type);
         } else {
@@ -142,18 +144,20 @@ $allMethodsContent
               name: 'body',
               isRequired: true,
               type: _getBodyParameterType(additionalParameter),
-              ref: additionalParameter.ref ?? additionalParameter.items?.ref));
+              ref: additionalParameter?.ref ??
+                  additionalParameter?.items?.ref ??
+                  ''));
         }
 
         final requestBody = swaggerRequest.requestBody?.ref;
-        if (requestBody != null) {
-          var requestBodyType = requestBody.split('/').last;
+        if (requestBody?.isNotEmpty == true) {
+          var requestBodyType = requestBody!.split('/').last;
 
           final bodySchema =
               requestBodies == null ? null : requestBodies[requestBodyType];
 
-          if (bodySchema != null) {
-            final content = bodySchema['content'] as Map;
+          if (bodySchema?.isNotEmpty == true) {
+            final content = bodySchema['content'] as Map?;
             final firstContent =
                 content == null ? null : content[content.keys.first];
             final schema = firstContent['schema'] as Map;
@@ -186,19 +190,19 @@ $allMethodsContent
         );
 
         final hasEnums = swaggerRequest.parameters.any((parameter) =>
-            parameter.items?.enumValues != null ||
-            parameter.item?.enumValues != null ||
-            parameter.schema?.enumValues != null ||
+            parameter.items?.enumValues.isNotEmpty == true ||
+            parameter.item?.enumValues.isNotEmpty == true ||
+            parameter.schema?.enumValues.isNotEmpty == true ||
             allEnumNames.contains(
-                'enums.${parameter.ref?.split("/")?.last?.pascalCase}'));
+                'enums.${parameter.ref.isEmpty ? "" : parameter.ref.split("/").last.pascalCase}'));
 
-        final enumInBodyName = swaggerRequest.parameters.firstWhere(
-            (parameter) =>
-                parameter.inParameter == 'body' &&
-                (parameter.items?.enumValues != null ||
-                    parameter.item?.enumValues != null ||
-                    parameter.schema?.enumValues != null),
-            orElse: () => null);
+        final enumInBodyName = swaggerRequest.parameters.firstWhereOrNull(
+          (parameter) =>
+              parameter.inParameter == 'body' &&
+              (parameter.items?.enumValues.isNotEmpty == true ||
+                  parameter.item?.enumValues.isNotEmpty == true ||
+                  parameter.schema?.enumValues.isNotEmpty == true),
+        );
 
         final parameterCommentsForMethod =
             getParameterCommentsForMethod(swaggerRequest.parameters, options);
@@ -221,7 +225,7 @@ $allMethodsContent
             hasFormData: hasFormData,
             returnType: returnTypeName,
             hasEnums: hasEnums,
-            enumInBodyName: enumInBodyName?.name,
+            enumInBodyName: enumInBodyName?.name ?? '',
             ignoreHeaders: options.ignoreHeaders,
             allEnumNames: allEnumNames,
             parameters: swaggerRequest.parameters);
@@ -233,17 +237,17 @@ $allMethodsContent
     return methods.toString();
   }
 
-  String _getBodyParameterType(RequestContent content) {
+  String _getBodyParameterType(RequestContent? content) {
     if (content == null) {
       return 'Object';
     }
 
-    if (content.type?.toLowerCase() == 'array') {
-      if (content.items.ref == null) {
+    if (content.type.toLowerCase() == 'array') {
+      if (content.items?.ref == null || content.items!.ref.isEmpty) {
         return 'Object';
       }
 
-      final type = content.items.ref.split('/').last.capitalize;
+      final type = content.items!.ref.split('/').last.capitalize;
 
       return 'List<$type>';
     }
@@ -271,7 +275,7 @@ $allMethodsContent
     if (inParameter == 'header' && options.ignoreHeaders) {
       return '';
     }
-    if (parameterDescription != null) {
+    if (parameterDescription.isNotEmpty) {
       parameterDescription =
           parameterDescription.replaceAll(RegExp(r'\n|\r|\t'), ' ');
     } else {
@@ -312,11 +316,11 @@ $allMethodsContent
         .toList();
 
     final enumParametersNames = parameters
-        .where((parameter) => (parameter.items?.enumValues != null ||
-            parameter.item?.enumValues != null ||
-            parameter.schema?.enumValues != null ||
+        .where((parameter) => (parameter.items?.enumValues.isNotEmpty == true ||
+            parameter.item?.enumValues.isNotEmpty == true ||
+            parameter.schema?.enumValues.isNotEmpty == true ||
             allEnumNames.contains(
-                'enums.${parameter.ref?.split("/")?.last?.pascalCase}')))
+                'enums.${parameter.ref.isNotEmpty ? parameter.ref.split("/").last.pascalCase : ""}')))
         .map((e) => e.name)
         .toList();
 
@@ -336,13 +340,13 @@ $allMethodsContent
 
   String getEnumParameter(String requestPath, String requestType,
       String parameterName, List<SwaggerRequestParameter> parameters,
-      [String ref]) {
+      [String ref = '']) {
     final enumListParametersNames = parameters
         .where((parameter) =>
             parameter.type == 'array' &&
-            (parameter.items?.enumValues != null ||
-                parameter.item?.enumValues != null ||
-                parameter.schema?.enumValues != null))
+            (parameter.items?.enumValues.isNotEmpty == true ||
+                parameter.item?.enumValues.isNotEmpty == true ||
+                parameter.schema?.enumValues.isNotEmpty == true))
         .map((e) => e.name)
         .toList();
 
@@ -356,7 +360,7 @@ $allMethodsContent
   }
 
   String validateParameterName(String parameterName) {
-    if (parameterName == null) {
+    if (parameterName.isEmpty) {
       return parameterName;
     }
 
@@ -378,35 +382,36 @@ $allMethodsContent
     return name.join();
   }
 
-  String getMethodContent(
-      {String summary,
-      String typeRequest,
-      String methodName,
-      String parametersContent,
-      String parametersComments,
-      String requestPath,
-      bool hasFormData,
-      String returnType,
-      bool hasEnums,
-      String enumInBodyName,
-      bool ignoreHeaders,
-      List<SwaggerRequestParameter> parameters,
-      List<String> allEnumNames}) {
+  String getMethodContent({
+    required String summary,
+    required String typeRequest,
+    required String methodName,
+    required String parametersContent,
+    required String parametersComments,
+    required String requestPath,
+    required bool hasFormData,
+    required String returnType,
+    required bool hasEnums,
+    required String enumInBodyName,
+    required bool ignoreHeaders,
+    required List<SwaggerRequestParameter> parameters,
+    required List<String> allEnumNames,
+  }) {
     var typeReq = typeRequest.capitalize + "(path: '$requestPath')";
     if (hasFormData) {
       typeReq +=
           '\n  @FactoryConverter(request: FormUrlEncodedConverter.requestFactory)';
     }
 
-    if (returnType != null && returnType != 'num') {
+    if (returnType.isNotEmpty && returnType != 'num') {
       returnType = returnType.pascalCase;
     }
 
-    final returnTypeString = returnType != null ? '<$returnType>' : '';
+    final returnTypeString = returnType.isNotEmpty ? '<$returnType>' : '';
     var parametersPart =
         parametersContent.isEmpty ? '' : '{$parametersContent}';
 
-    if (summary != null) {
+    if (summary.isNotEmpty) {
       summary = summary.replaceAll(RegExp(r'\n|\r|\t'), ' ');
     }
 
@@ -451,8 +456,8 @@ $allMethodsContent
   String validateParameterType(String parameterName) {
     var isEnum = false;
 
-    if (parameterName == null) {
-      return parameterName;
+    if (parameterName.isEmpty) {
+      return 'dynamic';
     }
 
     if (parameterName.startsWith('enums.')) {
@@ -473,7 +478,7 @@ $allMethodsContent
     }
   }
 
-  String getParameterTypeName(String parameter, [String itemsType]) {
+  String getParameterTypeName(String parameter, [String itemsType = '']) {
     switch (parameter) {
       case 'integer':
       case 'int':
@@ -491,22 +496,22 @@ $allMethodsContent
       case 'object':
         return 'Object';
       default:
-        return validateParameterType(parameter) ?? 'dynamic';
+        return validateParameterType(parameter);
     }
   }
 
   String getBodyParameter(SwaggerRequestParameter parameter, String path,
       String requestType, List<String> allEnumNames) {
     String parameterType;
-    if (parameter.type != null) {
+    if (parameter.type.isNotEmpty) {
       parameterType = parameter.type;
-    } else if (parameter.schema?.enumValues != null) {
+    } else if (parameter.schema?.enumValues.isNotEmpty ?? false) {
       parameterType =
           'enums.${SwaggerModelsGenerator.generateRequestEnumName(path, requestType, parameter.name)}';
-    } else if (parameter.schema?.originalRef != null) {
+    } else if (parameter.schema?.originalRef.isNotEmpty ?? false) {
       parameterType = SwaggerModelsGenerator.getValidatedClassName(
-          parameter.schema.originalRef.toString());
-    } else if (parameter.ref != null) {
+          parameter.schema!.originalRef.toString());
+    } else if (parameter.ref.isNotEmpty) {
       parameterType = parameter.ref.split('/').last;
       parameterType = parameterType.split('_').map((e) => e.capitalize).join();
 
@@ -517,8 +522,8 @@ $allMethodsContent
       if (parameter.type == 'array') {
         parameterType = 'List<$parameterType>';
       }
-    } else if (parameter.schema?.ref != null) {
-      parameterType = parameter.schema.ref.split('/').last;
+    } else if (parameter.schema?.ref.isNotEmpty ?? false) {
+      parameterType = parameter.schema!.ref.split('/').last;
     } else {
       parameterType = defaultBodyParameter;
     }
@@ -531,34 +536,37 @@ $allMethodsContent
   String getDefaultParameter(
       SwaggerRequestParameter parameter, String path, String requestType) {
     String parameterType;
-    if (parameter.schema?.enumValues != null) {
+    if (parameter.schema?.enumValues.isNotEmpty ?? false) {
       parameterType =
           'enums.${SwaggerModelsGenerator.generateRequestEnumName(path, requestType, parameter.name)}';
-    } else if (parameter.items?.enumValues != null) {
+    } else if (parameter.items?.enumValues.isNotEmpty ?? false) {
       final typeName =
           'enums.${SwaggerModelsGenerator.generateRequestEnumName(path, requestType, parameter.name)}';
       parameterType = 'List<$typeName>';
     } else {
-      parameterType = getParameterTypeName(
-          parameter.type ?? parameter.schema?.type, parameter.items?.type);
+      final neededType =
+          parameter.type.isNotEmpty ? parameter.type : parameter.schema!.type;
+
+      parameterType =
+          getParameterTypeName(neededType, parameter.items?.type ?? '');
     }
 
     return "@${parameter.inParameter.capitalize}('${parameter.name}') ${parameter.isRequired ? "@required" : ""} $parameterType? ${validateParameterName(parameter.name)}";
   }
 
   String getParameterContent(
-      {SwaggerRequestParameter parameter,
-      bool ignoreHeaders,
-      String requestType,
-      String path,
-      List<String> allEnumNames,
-      bool useRequiredAttribute}) {
+      {required SwaggerRequestParameter parameter,
+      required bool ignoreHeaders,
+      required String requestType,
+      required String path,
+      required List<String> allEnumNames,
+      required bool useRequiredAttribute}) {
     final parameterType = validateParameterType(parameter.name);
     switch (parameter.inParameter) {
       case 'body':
         return getBodyParameter(parameter, path, requestType, allEnumNames);
       case 'formData':
-        final isEnum = parameter.schema?.enumValues != null;
+        final isEnum = parameter.schema?.enumValues.isNotEmpty ?? false;
 
         return "@Field('${parameter.name}') ${parameter.isRequired ? "@required" : ""} ${isEnum ? 'enums.$parameterType?' : getParameterTypeName(parameter.type)}? ${validateParameterName(parameter.name)}";
       case 'header':
@@ -566,7 +574,7 @@ $allMethodsContent
             parameter.isRequired && useRequiredAttribute;
         return ignoreHeaders
             ? ''
-            : "@${parameter.inParameter.capitalize}('${parameter.name}') ${needRequiredAttribute ? "required" : ""} String?${validateParameterName(parameter.name)}";
+            : "@${parameter.inParameter.capitalize}('${parameter.name}') ${needRequiredAttribute ? "required" : ""} String? ${validateParameterName(parameter.name)}";
       case 'cookie':
         return '';
       default:
@@ -577,7 +585,7 @@ $allMethodsContent
   String getChopperClientContent(String fileName, String host, String basePath,
       GeneratorOptions options, bool hadModels) {
     final baseUrlString = options.withBaseUrl
-        ? "baseUrl:  'https://$host${basePath ?? ''}'"
+        ? "baseUrl:  'https://$host$basePath'"
         : '/*baseUrl: YOUR_BASE_URL*/';
 
     final converterString =
@@ -612,12 +620,12 @@ abstract class $className extends ChopperService''';
   }
 
   String getAllParametersContent(
-      {List<SwaggerRequestParameter> listParameters,
-      bool ignoreHeaders,
-      String path,
-      String requestType,
-      List<String> allEnumNames,
-      bool useRequiredAttribute}) {
+      {required List<SwaggerRequestParameter> listParameters,
+      required bool ignoreHeaders,
+      required String path,
+      required String requestType,
+      required List<String> allEnumNames,
+      required bool useRequiredAttribute}) {
     return listParameters
         .map((SwaggerRequestParameter parameter) => getParameterContent(
               parameter: parameter,
@@ -631,12 +639,19 @@ abstract class $className extends ChopperService''';
         .join(', ');
   }
 
-  SwaggerResponse getSuccessedResponse(List<SwaggerResponse> responses) {
-    return responses.firstWhere(
-        (SwaggerResponse response) =>
-            successDescriptions.contains(response.description) ||
-            successResponseCodes.contains(response.code),
-        orElse: () => null);
+  SwaggerResponse? getSuccessedResponse(List<SwaggerResponse> responses) {
+    return responses.firstWhereOrNull(
+      (SwaggerResponse response) =>
+          successDescriptions.contains(response.description) ||
+          successResponseCodes.contains(response.code),
+    );
+  }
+
+  String getResponseModelName(String url, String methodName) {
+    final urlString = url.split('/').map((e) => e.pascalCase).join();
+    final methodNamePart = methodName.pascalCase;
+
+    return '$urlString$methodNamePart\$Response';
   }
 
   String getReturnTypeName(
@@ -647,7 +662,7 @@ abstract class $className extends ChopperService''';
       List<String> dynamicResponses,
       Map<String, String> basicTypesMap) {
     if (overriddenRequests
-            ?.any((ResponseOverrideValueMap element) => element.url == url) ==
+            .any((ResponseOverrideValueMap element) => element.url == url) ==
         true) {
       final overriddenResponse = overriddenRequests
           .firstWhere((ResponseOverrideValueMap element) => element.url == url);
@@ -660,22 +675,28 @@ abstract class $className extends ChopperService''';
     final neededResponse = getSuccessedResponse(responses);
 
     if (neededResponse == null) {
-      return null;
+      return '';
     }
 
-    if (neededResponse.schema?.type != null) {
-      return getParameterTypeName(
-          neededResponse.schema.type,
-          neededResponse.schema?.items?.originalRef ??
-              neededResponse.schema?.items?.type ??
-              neededResponse.schema.items?.ref?.split('/')?.last);
+    if (neededResponse.schema?.type == 'object' &&
+        neededResponse.schema?.properties.isNotEmpty == true) {
+      return getResponseModelName(url, methodName);
     }
 
-    if (neededResponse.schema?.ref != null) {
-      return neededResponse.schema?.ref?.split('/')?.last;
+    if (neededResponse.schema?.type.isNotEmpty ?? false) {
+      final param = neededResponse.schema?.items?.originalRef.isNotEmpty == true
+          ? neededResponse.schema?.items?.originalRef
+          : neededResponse.schema?.items?.type.isNotEmpty == true
+              ? neededResponse.schema?.items?.type
+              : neededResponse.schema?.items?.ref.split('/').lastOrNull ?? '';
+      return getParameterTypeName(neededResponse.schema?.type ?? '', param!);
     }
 
-    if (neededResponse.ref != null) {
+    if (neededResponse.schema?.ref.isNotEmpty ?? false) {
+      return neededResponse.schema?.ref.split('/').lastOrNull ?? '';
+    }
+
+    if (neededResponse.ref.isNotEmpty) {
       final ref = neededResponse.ref.split('/').last.capitalize;
 
       if (neededResponse.ref.contains('/responses') &&
@@ -686,36 +707,42 @@ abstract class $className extends ChopperService''';
       }
     }
 
-    if (neededResponse.schema?.originalRef != null) {
-      return neededResponse.schema.originalRef;
+    if (neededResponse.schema?.originalRef.isNotEmpty ?? false) {
+      return neededResponse.schema?.originalRef ?? '';
     }
 
-    if (neededResponse.content?.isNotEmpty == true &&
-        neededResponse.content != null) {
-      if (neededResponse.content.first.ref != null) {
+    if (neededResponse.content.isNotEmpty == true &&
+        neededResponse.content.isNotEmpty) {
+      if (neededResponse.content.first.ref.isNotEmpty) {
         final type = neededResponse.content.first.ref.split('/').last;
 
         if (basicTypesMap.containsKey(type.toString())) {
-          return basicTypesMap[type];
+          return basicTypesMap[type] ?? '';
         }
         return type;
       }
-      if (neededResponse.content.first.responseType?.isNotEmpty == true) {
+      if (neededResponse.content.first.responseType.isNotEmpty) {
         return getParameterTypeName(
             neededResponse.content.first.responseType,
             neededResponse.schema?.items?.originalRef ??
-                neededResponse.content?.first?.items?.ref?.split('/')?.last);
+                neededResponse.content.firstOrNull?.items?.ref
+                    .split('/')
+                    .lastOrNull ??
+                '');
       }
     }
 
-    return null;
+    return '';
   }
 
   String getMapName(
       String path, String requestType, String parameterName, String ref) {
-    final enumName = ref?.split('/')?.last?.pascalCase ??
-        SwaggerModelsGenerator.generateRequestEnumName(
-            path, requestType, parameterName);
+    if (ref.isNotEmpty) {
+      return 'enums.\$${ref.split('/').lastOrNull?.pascalCase}Map';
+    }
+
+    final enumName = SwaggerModelsGenerator.generateRequestEnumName(
+        path, requestType, parameterName);
 
     return 'enums.\$${enumName}Map';
   }
