@@ -6,20 +6,20 @@ class SwaggerEnumsGeneratorV3 extends SwaggerEnumsGenerator {
   String generate(String dartCode, String fileName) {
     final dynamic map = jsonDecode(dartCode);
 
-    final components = map['components'] as Map<String, dynamic>;
+    final components = map['components'] as Map<String, dynamic>?;
     final schemas = components == null
         ? null
-        : components['schemas'] as Map<String, dynamic>;
+        : components['schemas'] as Map<String, dynamic>?;
 
     final responses = components == null
         ? null
-        : components['responses'] as Map<String, dynamic>;
+        : components['responses'] as Map<String, dynamic>?;
 
     final requestBodies = components == null
         ? null
-        : components['requestBodies'] as Map<String, dynamic>;
+        : components['requestBodies'] as Map<String, dynamic>?;
 
-    return generateFromMap(
-        dartCode, fileName, schemas, responses, requestBodies);
+    return generateFromMap(dartCode, fileName, schemas ?? {}, responses ?? {},
+        requestBodies ?? {});
   }
 }
