@@ -4,7 +4,8 @@ part 'parameter_item.g2.dart';
 
 @JsonSerializable()
 class ParameterItem {
-  ParameterItem({this.type, this.enumValues, this.defaultValue});
+  ParameterItem(
+      {this.type = '', this.enumValues = const [], this.defaultValue = ''});
 
   factory ParameterItem.fromJson(Map<String, dynamic> json) =>
       _$ParameterItemFromJson(json);
