@@ -8,15 +8,16 @@ part of 'response_schema.dart';
 
 ResponseSchema _$ResponseSchemaFromJson(Map<String, dynamic> json) {
   return ResponseSchema(
-    type: json['type'] as String,
-    ref: json['\$ref'] as String,
-    originalRef: json['originalRef'] as String,
-    enumValue: (json['enum'] as List<dynamic>)
-        ?.map((dynamic e) => e as String)
-        ?.toList(),
+    type: json['type'] as String? ?? '',
+    originalRef: json['originalRef'] as String? ?? '',
+    enumValue:
+        (json['enum'] as List<dynamic>?)?.map((e) => e as String).toList() ??
+            [],
+    properties: json['properties'] as Map<String, dynamic>? ?? {},
     items: json['items'] == null
         ? null
         : ItemSchema.fromJson(json['items'] as Map<String, dynamic>),
+    ref: json[r'$ref'] as String? ?? '',
   );
 }
 
@@ -24,7 +25,8 @@ Map<String, dynamic> _$ResponseSchemaToJson(ResponseSchema instance) =>
     <String, dynamic>{
       'type': instance.type,
       'originalRef': instance.originalRef,
+      r'$ref': instance.ref,
       'enum': instance.enumValue,
-      '\$ref': instance.ref,
       'items': instance.items,
+      'properties': instance.properties,
     };
